BE/requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
databricks-cli==0.18.0
duckdb==1.2.2
fastapi==0.119.0
gunicorn==22.0.0
json5==0.9.15                # latest stable version compatible with Py3.12
jsonpointer==2.1
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
polars==1.27.0
protobuf==4.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==1.0.1
# pywin32 removed — Windows-only, not needed on Linux/Mac
# PyQt5 removed — can be installed separately if GUI needed
# PyQtWebEngine removed — same as above
# pyls-spyder removed — optional; only for Spyder IDE
starlette==0.48.0
uvicorn==0.38.0
wheel==0.44.0
matplotlib==3.9.2
# mkl-service removed — only needed if using Intel MKL optimizations

#JWT
python-jose[cryptography]==3.5.0
requests==2.32.5
cryptography==46.0.3


